+++ app/controllers/likes_controller.rb
@@ -1,7 +1,6 @@
 class LikesController < ApplicationController
-    
+
     before_action :authenticate_user!
-    
     def toggle_like
         @post = Post.find(params[:post_id])
         like = Like.where(user: current_user, post: @post).first
@@ -11,19 +10,22 @@ class LikesController < ApplicationController
             like.destroy!
             @is_liked = false
         else
-            Like.create(user: current_user, post: @post)
-            @is_liked = true
-            if dislike
-                dislike.destroy!
-                @is_disliked = false
+            if current_user.authorized(@post)
+                Like.create(user: current_user, post: @post)
+                @is_liked = true
+                if dislike
+                    dislike.destroy!
+                    @is_disliked = false
+                end
             end
-        
+
         end
-        
+
         respond_to do |format|
             format.js { }
         end
-    
+
     end
+
 
 end
